requirements.txt
mlflow[databricks]>=3.1.0
langgraph==0.3.4
uv==0.6.10
databricks-sdk
databricks-agents
databricks-langchain
databricks-vectorsearch
unitycatalog-langchain
databricks-sqlalchemy
langchain-openai
langchain
langchain_core
langchain_community
lark
langgraph-supervisor